+++ LocalPayment/src/main/java/com/braintreepayments/api/localpayment/LocalPaymentTokenizeCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.localpayment
 
 /**
  * Callback for receiving result of [LocalPaymentClient.tokenize].