+++ LocalPayment/src/test/java/com/braintreepayments/api/localpayment/LocalPaymentApiUnitTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.localpayment;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNotNull;
@@ -13,8 +13,16 @@ import static org.mockito.Mockito.verify;
 
 import android.net.Uri;
 
+import com.braintreepayments.api.Fixtures;
+import com.braintreepayments.api.MockBraintreeClientBuilder;
 import com.braintreepayments.api.core.BraintreeClient;
 import com.braintreepayments.api.core.PostalAddress;
+import com.braintreepayments.api.localpayment.LocalPaymentApi;
+import com.braintreepayments.api.localpayment.LocalPaymentAuthRequestParams;
+import com.braintreepayments.api.localpayment.LocalPaymentInternalAuthRequestCallback;
+import com.braintreepayments.api.localpayment.LocalPaymentInternalTokenizeCallback;
+import com.braintreepayments.api.localpayment.LocalPaymentNonce;
+import com.braintreepayments.api.localpayment.LocalPaymentRequest;
 import com.braintreepayments.api.sharedutils.HttpResponseCallback;
 
 import org.json.JSONException;